--- crates/typst_element/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "typst_element"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-readme.workspace = true
-
-[dependencies]
-typst = { workspace = true }
-unicode-math-class = { workspace = true }
-paste = { workspace = true }
-
-[lints]
-workspace = true
