+++ crates/velyst/src/lib.rs
@@ -9,11 +9,11 @@ use world::VelystWorldPlugin;
 pub use imaging;
 pub use kanva;
 pub use typst;
-pub use typst_element;
 
 pub mod prelude {
     pub use crate::VelystSet;
     pub use crate::asset::{VelystModules, VelystSource};
+    pub use crate::ext::{ScopeError, ScopeExt, UnitExt};
     pub use crate::func::{
         TypstFunc, TypstFuncAppExt, TypstValue, VelystContent,
         VelystFunc, VelystSourceReady,
@@ -23,10 +23,14 @@ pub mod prelude {
     };
     pub use crate::typst_func;
     pub use crate::world::VelystWorld;
-    pub use typst_element::prelude::*;
+    pub use typst::foundations::Content;
+    pub use typst::foundations::NativeElement;
+    pub use typst::layout::Abs;
+    pub use typst::visualize as viz;
 }
 
 pub mod asset;
+pub mod ext;
 pub mod func;
 pub mod renderer;
 pub mod world;
